File structure.
