+++ src/angular/form-elements/validation/model/index.ts
@@ -1,0 +1,2 @@
+export * from "./validation.type";
+export * from "./validation-control.model";
